Replace query-editor cra with vite
