--- query-editor/src/JsonEditor.d.ts
@@ -1,3 +1,0 @@
-declare module "react-jsondata-editor" {
-  const JsonEditor: FunctionComponent<Props>;
-}
